# Conflicts:
#	core/assets/bundles/bundle_ja.properties
#	core/assets/bundles/bundle_zh_CN.properties
#	core/src/mindustry/client/ui/SchematicBrowserDialog.java
#	core/src/mindustry/entities/comp/UnitComp.java
#	core/src/mindustry/graphics/MultiPacker.java
#	core/src/mindustry/input/InputHandler.java
#	core/src/mindustry/mod/Mods.java

+++ core/src/mindustry/ui/dialogs/SettingsMenuDialog.java
@@ -34,7 +34,7 @@ import mindustry.logic.*;
 import mindustry.service.*;
 import mindustry.type.*;
 import mindustry.ui.*;
-import mindustry.ui.fragments.ChatFragment;
+import mindustry.ui.fragments.*;
 import mindustry.world.blocks.*;
 import mindustry.world.blocks.distribution.*;
 import mindustry.world.blocks.power.*;
@@ -149,6 +149,7 @@ public class SettingsMenuDialog extends BaseDialog{
 
             t.button("@settings.clearcampaignsaves", Icon.trash, style, () -> {
                 ui.showConfirm("@confirm", "@settings.clearcampaignsaves.confirm", () -> {
+                    control.saves.load(true);
                     for(var planet : content.planets()){
                         for(var sec : planet.sectors){
                             sec.clearInfo();
@@ -159,12 +160,14 @@ public class SettingsMenuDialog extends BaseDialog{
                         }
                     }
 
-                    control.saves.load(true);
-                    for(var slot : control.saves.getSaveSlots().copy()){
-                        if(slot.isSector()){
-                            slot.delete();
+                    if (control.saves.loadedSaveCount() > 0){ // If we don't explicitly check for loaded saves, getSaveSlots() will load *all* saves when there are 0 sectors loaded which we don't want as it's needlessly slow.
+                        for(var slot : control.saves.getSaveSlots().copy()){
+                            if(slot.isSector()){
+                                slot.delete();
+                            }
                         }
                     }
+                    control.saves.unload();
                 });
             }).marginLeft(4);
 
@@ -376,12 +379,12 @@ public class SettingsMenuDialog extends BaseDialog{
         client.sliderPref("minzoom", 0, 0, 100, s -> Strings.fixed(Mathf.pow(10, 0.0217f * s) / 100f, 2) + "x");
         client.sliderPref("weatheropacity", 50, 0, 100, s -> s + "%");
         client.sliderPref("beamdrillopacity", 100, 0, 100, 1, s -> s + "%");
-        client.sliderPref("junctionview", 0, -1, 1, 1, s -> { Junction.setBaseOffset(s); return s == -1 ? "On left side" : s == 1 ? "On right side" : "Do not show"; });
-        client.sliderPref("spawntime", 5, -1, 60, s -> { ClientVars.spawnTime = 60 * s; if (Vars.pathfinder.thread == null) Vars.pathfinder.start(); return s == -1 ? "Solid Line" : s == 0 ? "Disabled" : String.valueOf(s); });
-        client.sliderPref("traveltime", 10, 0, 60, s -> { ClientVars.travelTime = 60f / s; return s == 0 ? "Disabled" : String.valueOf(s); });
+        client.sliderPref("junctionview", 0, -1, 1, 1, s -> { Junction.setBaseOffset(s); return s == -1 ? "@client.left" : s == 1 ? "@client.right" : "Do not show"; });
+        client.sliderPref("spawntime", 5, -1, 60, s -> { ClientVars.spawnTime = 60 * s; if (Vars.pathfinder.thread == null) Vars.pathfinder.start(); return s == -1 ? "Solid Line" : s == 0 ? "@off" : String.valueOf(s); });
+        client.sliderPref("traveltime", 10, 0, 60, s -> { ClientVars.travelTime = 60f / s; return s == 0 ? "@off" : String.valueOf(s); });
         client.sliderPref("formationopacity", 30, 10, 100, 5, s -> { UnitType.formationAlpha = s / 100f; return s + "%"; });
-        client.sliderPref("hitboxopacity", 0, 0, 100, 5, s -> { UnitType.hitboxAlpha = s / 100f; return s == 0 ? "Disabled" : s + "%"; });
-        client.sliderPref("transferrangeopacity", 30, 0, 100, 5, s -> s + "%");
+        client.sliderPref("hitboxopacity", 0, 0, 100, 5, s -> { UnitType.hitboxAlpha = s / 100f; return s == 0 ? "@off" : s + "%"; });
+        client.sliderPref("transferrangeopacity", 30, 0, 100, 5, s -> s == 0 ? "@off" : s + "%");
         client.checkPref("tilehud", true);
         client.checkPref("lighting", true);
         client.checkPref("placementfragmentsearch", true);
@@ -413,6 +416,7 @@ public class SettingsMenuDialog extends BaseDialog{
         client.checkPref("useiconslogs", false);
         client.checkPref("colorizelogs", false);
         client.checkPref("showmassdriverdistance", false);
+        client.checkPref("enableunderwaterenv", false);
 
         client.category("misc");
         client.updatePref();
@@ -422,12 +426,12 @@ public class SettingsMenuDialog extends BaseDialog{
         client.textPref("keybind1command", "");
         client.sliderPref("minepathcap", 5000, -100, 5000, 100, s -> s == 0 ? "Unlimited" : s == -100 ? "Never" : String.valueOf(s));
         client.sliderPref("defaultbuildpathradius", 0, 0, 250, 5, s -> s == 0 ? "Unlimited" : String.valueOf(s));
-        client.sliderPref("modautoupdate", 1, 0, 2, s -> s == 0 ? "Disabled" : s == 1 ? "In Background" : "Restart Game");
+        client.sliderPref("modautoupdate", 1, 0, 2, s -> s == 0 ? "@off" : s == 1 ? "In Background" : "Restart Game");
         client.sliderPref("processorstatementscale", 80, 10, 100, 1, s -> String.format("%.2fx", s/100f)); // This is the most scuffed setting you have ever seen
         client.sliderPref("automapvote", 0, 0, 4, s -> s == 0 ? "Never" : s == 4 ? "Random vote" : "Always " + new String[]{"downvote", "novote", "upvote"}[--s]);
         client.sliderPref("pingexecutorthreads", OS.isWindows && !OS.is64Bit ? 5 : 65, 5, 105, 5, s -> s > 100 ? "Unlimited" : String.valueOf(s));
         client.sliderPref("maxschematicslisted", 300, 0, 3000, 150, s -> s == 0 ? "Unlimited" : String.valueOf(s));
-        client.textPref("defaultbuildpathargs", "self"); // Keep it to just self. Skill issue players going afk make this too problematic otherwise.
+        client.textPref("defaultbuildpathargs", "self"); // Keep it to just self. Skill issue players going afk make this too problematic otherwise. FINISHME: Add an afk detection system and revert this once we can reliably detect afk players and allow others to stop their pathing
         client.textPref("defaultminepathargs", "all");
         client.textPref("gamejointext", "");
         client.textPref("gamewintext", "");
@@ -457,7 +461,7 @@ public class SettingsMenuDialog extends BaseDialog{
         client.checkPref("schematicbrowserimporttags", true);
 
         if (settings.getBool("client-experimentals") || OS.hasProp("policone")) {
-            client.category("Experimental");
+            client.category("experimental");
             client.checkPref("trackcoreitems", false, i -> CoreItemsDisplay.trackItems = i && !net.server());
 
             client.checkPref("seer-warnings", false);
@@ -808,12 +812,27 @@ public class SettingsMenuDialog extends BaseDialog{
         protected Seq<Setting> list = new Seq<>();
         private static Seq<Setting> listSorted = new Seq<>();
         private String search = "";
-        private Table searchBarTable;
+        private final Table searchBarTable = makeSearchBar();
         private TextField searchBar;
+
+        private static final SettingsTable tempTable = new SettingsTable();
+
+        static {
+            tempTable.canRebuild = false; // Disable the hack that detects mods adding unsupported settings
+        }
 
         public SettingsTable(){
             left();
-            makeSearchBar();
+        }
+
+        private Table makeSearchBar(){
+            return new Table(s -> {
+                s.image(Icon.zoom);
+                searchBar = s.field(search, res -> {
+                    search = res;
+                    rebuild();
+                }).growX().get();
+            });
         }
 
         public Seq<Setting> getSettings(){
@@ -823,17 +842,6 @@ public class SettingsMenuDialog extends BaseDialog{
         public void pref(Setting setting){
             list.add(setting);
             rebuild();
-        }
-
-        private void makeSearchBar(){
-            searchBarTable = table(s -> {
-                s.image(Icon.zoom);
-                searchBar = s.field(search, res -> {
-                    search = res;
-                    rebuild();
-                }).growX().get();
-            }).get();
-            row();
         }
 
         public SliderSetting sliderPref(String name, int def, int min, int max, StringProcessor s){
@@ -885,16 +893,56 @@ public class SettingsMenuDialog extends BaseDialog{
         }
 
         private long lastRebuild;
-        public void rebuild(){
-            if (lastRebuild == -1) return; // Can't run more than twice per frame
-            if (lastRebuild == 0 || lastRebuild == Core.graphics.getFrameId()) { // First ever run and second run per frame
-                lastRebuild = -1;
-                Core.app.post(() -> {
-                    lastRebuild = -2; // Allows rebuild to run
-                    rebuild(); // This will set lastRebuild to the current frame
-                });
-                return;
+        /** True while in rebuild(). If the add() method is called on the settings table while this is false, we know that a mod is doing something unsupported and will disable the search bar */
+        private boolean isRebuilding;
+        /** Whether this table is safe to rebuild. A table is safe unless a mod has added a non Setting element. When this is false, rebuild() will always return immediately, even when called from the mod this table was created by */
+        private boolean canRebuild = true;
+        /** Horrible way to force a rebuild right before a mod adds a non Setting */
+        private boolean forceRebuild = false;
+
+        /** Scuffed way of detecting mods adding non Setting elements to tables */
+        @Override
+        public <T extends Element> Cell<T> add(T element){
+            // FINISHME: Do we also want to maybe get the stacktrace and check if this is being called from Setting.add()? It would be slow but some mod probably does something weird like this somewhere. This suggestion applies to row() as well
+            if(isRebuilding || !canRebuild) return super.add(element); // Normal behavior as long as we're adding stuff using the rebuild() method or once we've already disabled our optimizations
+            // Force the table to be rebuilt *before* adding the first non setting as this is how vanilla would behave (there's a good chance that the last rebuild was delayed to the next frame for performance reasons)
+            Log.warn(bundle.format("client.settings.search.disabled.log.add", Threads.getTrace(0)));
+            forceRebuild = true;
+            rebuild();
+            return super.add(element);
+        }
+
+        /** Scuffed way of detecting mods trying to add rows before we have rebuilt() at least once (this happens since foo's normally delays the first rebuild to the frame after setting creation) */
+        @Override
+        public Table row(){
+            if (hasChildren() || !canRebuild || isRebuilding) return super.row();
+            // If we haven't yet seen a non setting, but they're trying to add a row, they're probably just adding it after the Reset to Defaults button which would be here normally. Trigger a rebuild so its there properly
+            Log.warn(bundle.format("client.settings.search.disabled.log.row", Threads.getTrace(0)));
+            forceRebuild = true;
+            rebuild();
+            return super.row();
+        }
+
+        public void rebuild(){ // FINISHME: Ideally, we should still be able to search for settings even if a mod has added non settings to the table. Perhaps we could have a Setting type that just wraps a runnable that adds the element being added in add()?
+            if(!canRebuild) return;
+            if(!forceRebuild){
+                // This optimization breaks some mods that add content to the table as non Settings, it is automatically bypassed rebuilding() is blocked entirely when non Settings are detected
+                if (lastRebuild == -1) return; // Can't run more than twice per frame
+                if (lastRebuild == 0 || lastRebuild == Core.graphics.getFrameId()) { // First ever run and second run per frame
+                    lastRebuild = -1;
+                    Core.app.post(() -> {
+                        lastRebuild = -2; // Allows rebuild to run
+                        rebuild(); // This will set lastRebuild to the current frame
+                    });
+                    return;
+                }
+            }else{ // A rebuild was forced, this means that we can no longer rebuild this due to a mod adding a non Setting
+                canRebuild = false;
+                searchBar.setDisabled(true);
+                searchBar.addListener(Tooltip.Tooltips.getInstance().create("@client.settings.search.disabled.tooltip"));
             }
+
+            isRebuilding = true;
             lastRebuild = Core.graphics.getFrameId();
             boolean hasFocus = searchBar.hasKeyboard();
             clearChildren();
@@ -902,10 +950,26 @@ public class SettingsMenuDialog extends BaseDialog{
             add(searchBarTable).fillX().padBottom(4);
             row();
 
-            if(search.trim().isEmpty()){
+            if(search.trim().isEmpty()){ // No search value: Contains hacky code for client setting categories
+                Cons<Category> pushCategory = cat -> { // Adds all the elements of the current category to the settings table
+                    if(cat != null){
+                        cat.children.clearChildren(); // Clear the settings otherwise we'll end up with multiple sets of settings
+                        tempTable.getCells().each(cell -> {
+                            cat.children.add(cell.get()).set(cell); // Hack to get the element of a cell, move it to another table and then copy over the properties to its new cell.
+                            if(cell.isEndRow()) cat.children.row(); // Extra hack because setting the cell to the old one doesn't account for rows
+                        });
+                    }
+                    tempTable.clearChildren();
+                };
+                Category lastCategory = null;
                 for(Setting setting : list){
-                    setting.add(this);
+                    if(setting instanceof Category c){
+                        pushCategory.get(lastCategory);
+                        lastCategory = c;
+                        setting.add(this);
+                    }else setting.add(lastCategory == null ? this : tempTable);
                 }
+                pushCategory.get(lastCategory); // Handle final category for table
             }else{
                 listSorted.selectFrom(list, s -> !(s instanceof Category));
                 var searchLower = search.toLowerCase();
@@ -915,7 +979,7 @@ public class SettingsMenuDialog extends BaseDialog{
 //                    var desc = u.description == null ? "" : Strings.stripColors(u.description).toLowerCase();
                     var weight = title.isEmpty() /*&& desc.isEmpty()*/ ? Float.POSITIVE_INFINITY : 0f;
 
-                    if (!title.isEmpty()) weight += ClientUtils.biasedLevenshtein(searchLower, title, false, true) / (Structs.count(title.split(" "), searchSplit::contains) + 1);
+                    if (!title.isEmpty()) weight += ClientUtils.biasedLevenshtein(searchLower, title, true, true) / (Structs.count(title.split(" "), searchSplit::contains) + 1); // Case-sensitive as a minor optimization so that search query is only lowercased once
 //                    Line below doesn't work great since a lot of the settings don't have descriptions
 //                    if (!desc.isEmpty()) weight += .5f * ClientUtils.biasedLevenshtein(searchLower, desc, false, true) / (Structs.count(desc.split(" "), searchSplit::contains) + 1);
 
@@ -938,6 +1002,7 @@ public class SettingsMenuDialog extends BaseDialog{
             if(hasFocus){
                 searchBar.requestKeyboard();
             }
+            isRebuilding = false;
         }
 
         public abstract static class Setting{
@@ -947,6 +1012,12 @@ public class SettingsMenuDialog extends BaseDialog{
 
             public Setting(String name){
                 this.name = name;
+                String fooKey = "client.setting." + name + ".name";
+                if(bundle.has(fooKey)){ // Client keys are in the format client.setting.name.__ as we name all client settings this way and its cleaner. We also don't support winkey as we don't need it.
+                    title = bundle.get(fooKey);
+                    description = bundle.getOrNull("client.setting." + name + ".description");
+                    return;
+                }
                 String winkey = "setting." + name + ".name.windows";
                 title = OS.isWindows && bundle.has(winkey) ? bundle.get(winkey) : bundle.get("setting." + name + ".name", name);
                 description = bundle.getOrNull("setting." + name + ".description");
@@ -1033,16 +1104,20 @@ public class SettingsMenuDialog extends BaseDialog{
 
         // Elements are actually added below
         public static class Category extends Setting{
+            protected Table children = new Table();
+            private final Collapser collapser = new Collapser(children, settings.getBool("settingscategory-" + name + "-enabled", true));
+
             Category(String name){
                 super(name);
-                this.name = name;
-                this.title = bundle.get("setting." + name + ".category");
+                title = bundle.get("client.setting." + name + ".category");
             }
 
             @Override
             public void add(SettingsTable table){
                 table.add("").row(); // Add a cell first as .row doesn't work if there are no cells in the current row.
-                table.add("[accent]" + title);
+                table.check("[accent]" + title, !collapser.isCollapsed(), b -> { collapser.setCollapsed(!b); settings.put("settingscategory-" + name + "-enabled", !b); });
+                table.row();
+                table.add(collapser).left();
                 table.row();
             }
         }
@@ -1055,8 +1130,6 @@ public class SettingsMenuDialog extends BaseDialog{
 
                 @Override
                 public void add(SettingsTable table) { // Update URL with update button FINISHME: Move this to TextPref when i decide im willing to spend 6 hours doing so
-                    name = "updateurl";
-                    title = bundle.get("setting." + name + ".name");
 
                     table.table(t -> {
                         t.button(Icon.refresh, Styles.settingTogglei, 32, () -> {
